fix: unifica o intervalo de reposicionamento numa unica fonte de verdade
O motor de alertas e a tela de pacientes usavam janelas DIFERENTES, e por um
fator de exatamente dois:

  motor (configuracao.py)      : alto 60min, medio 90min, baixo 120min
  tela (paciente_service.py)   : alto 2h,    medio 3h,    baixo 4h

Ou seja, a ficha informava "reposicionar a cada 2h" para um paciente de alto
risco enquanto o sistema alertava a cada 1h. Num parametro clinico, duas fontes
de verdade significam que pelo menos uma esta errada — e ninguem consegue saber
qual olhando a tela. Confirmado com o dono do produto que a janela correta e a
do MOTOR; a tela e que estava errada.

Ao corrigir, apareceu algo pior: `repositioningInterval` era editavel no
formulario, enviado ao backend e SILENCIOSAMENTE DESCARTADO — nem
`create_patient` nem `update_patient` liam o campo. Quem alterasse aquele
numero acreditaria ter mudado o protocolo do paciente sem ter mudado nada.

Agora:
- `intervalo_horas(perfil)` deriva de `config.janela_por_perfil`, a mesma
  configuracao que o motor usa para disparar;
- os schemas passam a float — com a janela do motor, o perfil medio da 1.5h e
  um int truncaria para 1h, recriando a divergencia;
- o campo no formulario vira somente-leitura, com o texto explicando que e
  definido pelo perfil de risco;
- GET /api/perfis-risco expoe o mapeamento para o formulario mostrar o valor
  correto ANTES de salvar. Sem esse endpoint, o frontend precisaria de uma
  copia da tabela — um terceiro mapa para divergir dos outros dois.

Verificado: a API devolve {"high":1.0,"medium":1.5,"low":2.0}, que e
exatamente 60/90/120 min do motor. Teste parametrizado trava a igualdade entre
as duas pontas.

+++ interface/schemas.py
@@ -62,7 +62,13 @@ class FrontendCreatePatient(BaseModel):
     room: str | None = None
     bed: str | None = None
     riskLevel: str
-    repositioningInterval: int | None = None
+    # Aceito por compatibilidade com clientes antigos, mas IGNORADO: o intervalo
+    # e derivado do perfil de risco (ver paciente_service.intervalo_horas), que
+    # e a mesma fonte usada pelo motor de alertas. Antes o formulario deixava
+    # editar este numero, mandava para o backend e o valor era descartado sem
+    # aviso — a pessoa achava ter mudado o protocolo do paciente e nao mudara
+    # nada.
+    repositioningInterval: float | None = None
     notes: str | None = None
 
     @field_validator("riskLevel")
@@ -88,7 +94,9 @@ class FrontendPatient(BaseModel):
     room: str | None = None
     bed: str | None = None
     riskLevel: str
-    repositioningInterval: int | None = None
+    # float: com a janela do motor, o perfil medio da 1.5h — um int truncaria
+    # para 1h e a tela voltaria a divergir do comportamento real.
+    repositioningInterval: float | None = None
     createdAt: str | None = None
     updatedAt: str | None = None
